Feat: KAN-44 마이페이지 활동 기록 API 연결
